+++ clang/include/clang/Basic/Diagnostic.h
@@ -401,9 +401,9 @@ private:
     /// modifications done through the command-line.
     struct DiagStatePoint {
       DiagState *State;
-      unsigned Offset;
+      uint64_t Offset;
 
-      DiagStatePoint(DiagState *State, unsigned Offset)
+      DiagStatePoint(DiagState *State, uint64_t Offset)
           : State(State), Offset(Offset) {}
     };
 
@@ -416,7 +416,7 @@ private:
       File *Parent = nullptr;
 
       /// The offset of this file within its parent.
-      unsigned ParentOffset = 0;
+      uint64_t ParentOffset = 0;
 
       /// Whether this file has any local (not imported from an AST file)
       /// diagnostic state transitions.
@@ -426,7 +426,7 @@ private:
       /// be at least one of these (the state on entry to the file).
       llvm::SmallVector<DiagStatePoint, 4> StateTransitions;
 
-      DiagState *lookup(unsigned Offset) const;
+      DiagState *lookup(uint64_t Offset) const;
     };
 
     /// The diagnostic states for each file.
